requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pandas==2.1.1
psycopg2-binary
python-dotenv==1.0.0
python-multipart
